+++ src/data-migrations/index.js
@@ -1,0 +1,10 @@
+export {
+	registerMigratorSettings,
+	registerResultsRenderer,
+	registerMigratorRenderer,
+	getMigratorSettingsComponent,
+	getResultsRenderer,
+	getMigratorRenderer
+} from './registry'
+
+export { default as registerOnReady } from './registerOnReady'