+++ src/index.js
@@ -1,0 +1,6 @@
+export * as api from './api'
+export * as components from './components'
+export * as dataMigrations from './data-migrations'
+export * as hooks from './hooks'
+export * as settings from './settings'
+export * as utils from './utils'