+++ src/settings/index.js
@@ -1,0 +1,2 @@
+export { default as createUseSettings } from './createUseSettings'
+export { default as AdminSettingsPage } from './AdminSettingsPage'