+++ src/utils/constants.js
@@ -1,0 +1,24 @@
+export const ALERTS_TIMEOUT = 10000
+
+export const MIGRATION_RESULTS_PER_PAGE = 100
+
+export const SHOW_CANCEL_ALL_MIGRATIONS_BUTTON = false
+
+export const SHOW_MIGRATION_ID_IN_PREVIOUS_MIGRATIONS_SELECT = false
+
+export const LOG_LEVELS = [
+	'emergency',
+	'alert',
+	'critical',
+	'error',
+	'warning',
+	'notice',
+	'info',
+	'debug',
+	'off'
+]
+
+export const MIGRATION_ACTION_SCHEDULER_QUEUE_MODE = [
+	'async',
+	'scheduled'
+]