+++ src/webpack/postcss.config.js
@@ -1,0 +1,5 @@
+const postcssPlugins = require( '@wordpress/postcss-plugins-preset' )
+
+module.exports = {
+	plugins: postcssPlugins
+}